requirements.txt
streamlit
python-dotenv
openai
heyoo
pygsheets
pathlib
langchain
langchain-openai
pandas
email-validator